+++ python/opensartriad/__init__.py
@@ -23,7 +23,7 @@ from .client import (
     SceneCollection,
 )
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 __all__ = [
     "Catalog",
     "Scene",